Allow setting an explicit data address (host, port)
When using FTP begind a firewalll, it is useful to be able to specify what port the remote server will connect to for data transfers. This way, you will be able to open these ports beforehand, instead of relying on a random port being used.

This adds a data_address tuple to the connect() method that allows you to specify a (host, port) combination for this purpose.